LayoutTests/platform/gtk/imported/w3c/web-platform-tests/accname/name/comp_host_language_label-expected.txt

AccName: Host Language Label Tests

Tests the #comp_host_language_label portions of the AccName Name Computation algorithm.

HTML input with value, alt, etc.


HTML input label/for

checkbox label
color label
date label
datetime-local label
email label
month label
number label
password label
radio label
range label
search label
tel label
textfield label
time label
url label
week label
HTML input label encapsulation

checkbox label
color label
date label
datetime-local label
email label
month label
number label
password label
radio label
range label
search label
tel label
textfield label
time label
url label
week label
HTML select

select label
select label
HTML img/picture


HTML fieldset/legend

fieldset legend label

HTML table/caption

table caption label
a	b	c
1	2	3
SVG title




PASS html: input[type=button]
PASS html: input[type=image]
PASS html: input[type=reset]
PASS html: input[type=submit]
PASS html: label[for] input[type=checkbox]
PASS html: label[for] input[type=color]
PASS html: label[for] input[type=date]
PASS html: label[for] input[type=datetime-local]
PASS html: label[for] input[type=email]
PASS html: label[for] input[type=month]
PASS html: label[for] input[type=number]
PASS html: label[for] input[type=password]
PASS html: label[for] input[type=radio]
PASS html: label[for] input[type=range]
PASS html: label[for] input[type=search]
PASS html: label[for] input[type=tel]
PASS html: label[for] input[type=text]
PASS html: label[for] input[type=time]
PASS html: label[for] input[type=url]
PASS html: label[for] input[type=week]
PASS html: label input[type=checkbox] encapsulation
PASS html: label input[type=color] encapsulation
FAIL html: label input[type=date] encapsulation assert_equals: <input type="date" data-expectedlabel="date label" data-testname="html: label input[type=date] encapsulation" class="ex"> expected "date label" but got "month / day / year date label"
FAIL html: label input[type=datetime-local] encapsulation assert_equals: <input type="datetime-local" data-expectedlabel="datetime-local label" data-testname="html: label input[type=datetime-local] encapsulation" class="ex"> expected "datetime-local label" but got "month / day / year , 12 : 30 PM datetime-local label"
PASS html: label input[type=email] encapsulation
FAIL html: label input[type=month] encapsulation assert_equals: <input type="month" data-expectedlabel="month label" data-testname="html: label input[type=month] encapsulation" class="ex"> expected "month label" but got "month year month label"
PASS html: label input[type=number] encapsulation
PASS html: label input[type=password] encapsulation
PASS html: label input[type=radio] encapsulation
FAIL html: label input[type=range] encapsulation assert_equals: <input type="range" data-expectedlabel="range label" data-testname="html: label input[type=range] encapsulation" class="ex"> expected "range label" but got "50 range label"
FAIL html: label input[type=search] encapsulation assert_equals: <input type="search" data-expectedlabel="search label" data-testname="html: label input[type=search] encapsulation" class="ex"> expected "search label" but got "cancel search label"
PASS html: label input[type=tel] encapsulation
PASS html: label[for] input[type=text] encapsulation
FAIL html: label input[type=time] encapsulation assert_equals: <input type="time" data-expectedlabel="time label" data-testname="html: label input[type=time] encapsulation" class="ex"> expected "time label" but got "12 : 30 PM time label"
PASS html: label input[type=url] encapsulation
PASS html: label input[type=week] encapsulation
PASS html: select for/id
FAIL html: select encapsulation assert_equals: <select data-expectedlabel="select label" data-testname="html: select encapsulation" class="ex">
    <option>foo</option>
  </select> expected "select label" but got "select label foo"
PASS html: img[alt] (non-empty)
PASS html: picture > img[alt] (non-empty)
FAIL html: fieldset > legend assert_equals: <fieldset data-expectedlabel="fieldset legend label" data-testname="html: fieldset > legend" class="ex">
  <legend>fieldset legend label</legend>
  <input type="text"><br>
</fieldset> expected "fieldset legend label" but got ""
PASS html: table > caption
PASS svg: circle > title
PASS svg: rect > title
PASS svg: polygon > title
PASS svg: g > title
